pass converter registry to data store
With this change users will control from where to get(load) converter which will be used in current instance of data store.
