+++ pom.xml
@@ -4,7 +4,7 @@
 	<groupId>com.google.code.twig</groupId>
 	<artifactId>twig-persist</artifactId>
 	<packaging>jar</packaging>
-	<version>2.0-beta6</version>
+	<version>2.0-beta7</version>
 
 	<properties>
 		<appengine.version>1.9.15</appengine.version>
